requirements.txt
codegen==1.0
dj-database-url==0.5.0
Django==2.0.13
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-tinymce==3.0.2
gunicorn==19.9.0
patterns==0.3
Pillow==7.1.2
psycopg2==2.8.3
pytz==2019.2
whitenoise==4.1.4
